Backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14,<4
django-cors-headers>=4.0,<5
python-dotenv>=1.0,<2
requests>=2.31,<3
PyJWT>=2.8,<3
psycopg2-binary>=2.9,<3
gunicorn>=21.2,<22
